requirements.txt
django
django-background-tasks
pyserial
mysqlclient
